Fix even more errors
Issues with optionals should hopefully now be resolved, unless I missed any. Major issues remaining are with rendering changes, removal of tool/weapon items, and inventory changes.

+++ src/main/java/meteordevelopment/meteorclient/systems/modules/movement/NoFall.java
@@ -141,7 +141,7 @@ public class NoFall extends Module {
         // Airplace mode
         if (mode.get() == Mode.AirPlace) {
             // Test if fall damage setting is valid
-            if (!airPlaceMode.get().test(mc.player.fallDistance)) return;
+            if (!airPlaceMode.get().test((float) mc.player.fallDistance)) return;
 
             // Center and place block
             if (anchor.get()) PlayerUtils.centerPlayer();
